+++ app/build.gradle
@@ -47,8 +47,11 @@ android {
 }
 
 dependencies {
+    def nav_version = "2.5.2"
+
     implementation "io.coil-kt:coil-compose:2.2.1"
 
+    implementation "androidx.navigation:navigation-compose:$nav_version"
     implementation "androidx.compose.ui:ui-text-google-fonts:$compose_ui_version"
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.5.1'
